move list-event to share
