+++ src/frontend/src/app/portal/daemonset/list-daemonset/list-daemonset.component.ts
@@ -12,7 +12,7 @@ import {
 import { ConfirmationDialogService } from '../../../shared/confirmation-dialog/confirmation-dialog.service';
 import { Subscription } from 'rxjs/Subscription';
 import { PublishDaemonSetTplComponent } from '../publish-tpl/publish-tpl.component';
-import { ListEventComponent } from '../list-event/list-event.component';
+import { ListEventComponent } from '../../../shared/list-event/list-event.component';
 import { ListPodComponent } from '../../../shared/list-pod/list-pod.component';
 import { TplDetailService } from '../../../shared/tpl-detail/tpl-detail.service';
 import { AuthService } from '../../../shared/auth/auth.service';
