--- src/frontend/src/app/portal/daemonset/list-event/list-event.scss
@@ -1,14 +1,0 @@
-.form-group-label-override {
-  font-size: 14px;
-  font-weight: 400;
-}
-
-.access-level-label {
-  font-size: 14px;
-  font-weight: 400;
-  margin-left: -4px;
-  margin-right: 12px;
-  top: -6px;
-  position: relative;
-}
-
