+++ src/frontend/src/app/portal/deployment/deployment.module.ts
@@ -6,7 +6,6 @@ import { CreateEditDeploymentComponent } from './create-edit-deployment/create-e
 import { ReactiveFormsModule } from '@angular/forms';
 import { CreateEditDeploymentTplComponent } from './create-edit-deploymenttpl/create-edit-deploymenttpl.component';
 import { PublishDeploymentTplComponent } from './publish-tpl/publish-tpl.component';
-import { ListEventComponent } from './list-event/list-event.component';
 import { DeploymentClient } from '../../shared/client/v1/kubernetes/deployment';
 import { PodClient } from '../../shared/client/v1/kubernetes/pod';
 import { DeploymentService } from '../../shared/client/v1/deployment.service';
@@ -40,7 +39,6 @@ import { LogClient } from '../../shared/client/v1/kubernetes/log';
     ListDeploymentComponent,
     CreateEditDeploymentTplComponent,
     PublishDeploymentTplComponent,
-    ListEventComponent
   ]
 })
