+++ src/frontend/src/app/portal/statefulset/statefulset.module.ts
@@ -1,7 +1,6 @@
 import { NgModule } from '@angular/core';
 import { SharedModule } from '../../shared/shared.module';
 import { ReactiveFormsModule } from '@angular/forms';
-import { ListEventComponent } from './list-event/list-event.component';
 import { PodClient } from '../../shared/client/v1/kubernetes/pod';
 import { ClusterService } from '../../shared/client/v1/cluster.service';
 import { PublicService } from '../../shared/client/v1/public.service';
@@ -39,7 +38,6 @@ import { CreateEditStatefulsettplComponent } from './create-edit-statefulsettpl/
     CreateEditStatefulsetComponent,
     ListStatefulsetComponent,
     PublishStatefulsetTplComponent,
-    ListEventComponent,
     CreateEditStatefulsettplComponent
 
   ]
