+++ src/frontend/src/app/shared/list-event/list-event.scss
